requirements.txt
aiogram>=3.4.0,<4.0.0
SQLAlchemy[asyncio]>=2.0.0,<3.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
